+++ NEWS.md
@@ -16,10 +16,16 @@
 
 * Increased accuracy for `check_convergence()` for *glmmTMB* models.
 
+* `r2()` and `r2_mcfadden()` now support beta-binomial (non-mixed) models from
+  package *glmmTMB*.
+
 ## Bug fixes
 
 * `check_outliers()` did not warn that no numeric variables were found when only
   the response variable was numeric, but all relevant predictors were not.
+
+* `check_collinearity()` did not work for glmmTMB models when zero-inflation
+  component was set to `~0`.
 
 # performance 0.12.4